requirements.txt
# PyTorch
torch==2.10.0+cu126
torchvision==0.25.0+cu126
timm==0.6.7

# Lightning + Hydra
lightning==2.6.1
hydra-core==1.3.0
hydra-colorlog==1.2.0
rich==14.3.3

# Common utilities
numpy==1.23.5
scipy==1.15.3
pillow==12.1.1
tqdm==4.67.3
pyyaml==6.0.3
setuptools==82.0.1
opencv-python==4.11.0.86
ffmpeg-python==0.2.0
einops==0.8.2
imageio==2.34.1
av==13.1.0
safetensors==0.7.0

# 3D
trimesh==4.11.3

# ML
transformers==5.3.0
wandb==0.25.1
fastapi==0.116.1
gradio==5.47.0
uvicorn==0.35.0
